Address feedback - move GenerateNameForExpression functionality from shared layer up to Workspaces layer
