+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Extensions/SemanticModelExtensions.cs
@@ -11,11 +11,6 @@ using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Utilities;
 using Roslyn.Utilities;
 
-#if !CODE_STYLE
-using Microsoft.CodeAnalysis;
-using Humanizer;
-#endif
-
 namespace Microsoft.CodeAnalysis.CSharp.Extensions
 {
     internal static partial class SemanticModelExtensions
@@ -157,164 +152,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             }
 
             return true;
-        }
-
-        /// <summary>
-        /// Given an argument node, tries to generate an appropriate name that can be used for that
-        /// argument.
-        /// </summary>
-        public static string GenerateNameForArgument(
-            this SemanticModel semanticModel, ArgumentSyntax argument, CancellationToken cancellationToken)
-        {
-            // If it named argument then we use the name provided.
-            if (argument.NameColon != null)
-            {
-                return argument.NameColon.Name.Identifier.ValueText;
-            }
-
-            return semanticModel.GenerateNameForExpression(
-                argument.Expression, capitalize: false, cancellationToken: cancellationToken);
-        }
-
-        public static string GenerateNameForArgument(
-            this SemanticModel semanticModel, AttributeArgumentSyntax argument, CancellationToken cancellationToken)
-        {
-            // If it named argument then we use the name provided.
-            if (argument.NameEquals != null)
-            {
-                return argument.NameEquals.Name.Identifier.ValueText;
-            }
-
-            return semanticModel.GenerateNameForExpression(
-                argument.Expression, capitalize: false, cancellationToken: cancellationToken);
-        }
-
-        /// <summary>
-        /// Given an expression node, tries to generate an appropriate name that can be used for
-        /// that expression. 
-        /// </summary>
-        public static string GenerateNameForExpression(
-            this SemanticModel semanticModel, ExpressionSyntax expression,
-            bool capitalize, CancellationToken cancellationToken)
-        {
-            // Try to find a usable name node that we can use to name the
-            // parameter.  If we have an expression that has a name as part of it
-            // then we try to use that part.
-            var current = expression;
-            while (true)
-            {
-                current = current.WalkDownParentheses();
-
-                if (current is IdentifierNameSyntax identifierName)
-                {
-                    return identifierName.Identifier.ValueText.ToCamelCase();
-                }
-                else if (current is MemberAccessExpressionSyntax memberAccess)
-                {
-                    return memberAccess.Name.Identifier.ValueText.ToCamelCase();
-                }
-                else if (current is MemberBindingExpressionSyntax memberBinding)
-                {
-                    return memberBinding.Name.Identifier.ValueText.ToCamelCase();
-                }
-                else if (current is ConditionalAccessExpressionSyntax conditionalAccess)
-                {
-                    current = conditionalAccess.WhenNotNull;
-                }
-                else if (current is CastExpressionSyntax castExpression)
-                {
-                    current = castExpression.Expression;
-                }
-                else if (current is DeclarationExpressionSyntax decl)
-                {
-                    if (!(decl.Designation is SingleVariableDesignationSyntax name))
-                    {
-                        break;
-                    }
-
-                    return name.Identifier.ValueText.ToCamelCase();
-                }
-                else if (current.Parent is ForEachStatementSyntax foreachStatement &&
-                         foreachStatement.Expression == expression)
-                {
-                    var name = foreachStatement.Identifier.ValueText.ToCamelCase();
-#if !CODE_STYLE
-                    name = name.Pluralize();
-#endif
-                    return name;
-                }
-                else
-                {
-                    break;
-                }
-            }
-
-            // there was nothing in the expression to signify a name.  If we're in an argument
-            // location, then try to choose a name based on the argument name.
-            var argumentName = TryGenerateNameForArgumentExpression(
-                semanticModel, expression, cancellationToken);
-            if (argumentName != null)
-            {
-                return capitalize ? argumentName.ToPascalCase() : argumentName.ToCamelCase();
-            }
-
-            // Otherwise, figure out the type of the expression and generate a name from that
-            // instead.
-            var info = semanticModel.GetTypeInfo(expression, cancellationToken);
-
-            // If we can't determine the type, then fallback to some placeholders.
-            var type = info.Type;
-
-            var parameterName = type.CreateParameterName(capitalize);
-#if !CODE_STYLE
-            if (Pluralize(semanticModel, type))
-            {
-                parameterName = parameterName.Pluralize();
-            }
-#endif
-            return parameterName;
-        }
-
-#if !CODE_STYLE
-        private static bool Pluralize(SemanticModel semanticModel, ITypeSymbol type)
-        {
-            if (type == null)
-                return false;
-
-            if (type.SpecialType == SpecialType.System_String)
-                return false;
-
-            var enumerableType = semanticModel.Compilation.IEnumerableOfTType();
-            return type.AllInterfaces.Any(i => i.OriginalDefinition.Equals(enumerableType));
-        }
-#endif
-
-        private static string TryGenerateNameForArgumentExpression(
-            SemanticModel semanticModel, ExpressionSyntax expression, CancellationToken cancellationToken)
-        {
-            var topExpression = expression.WalkUpParentheses();
-            if (topExpression.IsParentKind(SyntaxKind.Argument, out ArgumentSyntax argument))
-            {
-                if (argument.NameColon != null)
-                {
-                    return argument.NameColon.Name.Identifier.ValueText;
-                }
-
-                if (argument.Parent is BaseArgumentListSyntax argumentList)
-                {
-                    var index = argumentList.Arguments.IndexOf(argument);
-                    if (semanticModel.GetSymbolInfo(argumentList.Parent, cancellationToken).Symbol is IMethodSymbol member && index < member.Parameters.Length)
-                    {
-                        var parameter = member.Parameters[index];
-                        if (parameter.Type.OriginalDefinition.TypeKind != TypeKind.TypeParameter)
-                        {
-                            return parameter.Name;
-                        }
-                    }
-                }
-            }
-
-            return null;
         }
 
         public static ISet<INamespaceSymbol> GetUsingNamespacesInScope(this SemanticModel semanticModel, SyntaxNode location)
